Javascript Final Project
Javascript Final Project featuring a website for a fictional fashion
brand called “Trashique” created by me. I incorporated instagram api to
show instagram photos with the hashtags: “hobo-chic”, “trash”, “raw”,
and “grunge.” This is a critical design piece commenting on the
misappropriation of culture and common glamorization of society’s
underdogs in the high fashion industry.